% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comebineSeg.R
\name{combineSeg}
\alias{combineSeg}
\title{Combine two chromosomal segments into one}
\usage{
combineSeg(seq.dat, max.brk.len = 3e+06, ap.calc = FALSE)
}
\arguments{
\item{seq.dat}{a data.frame containing the sequencing data}

\item{max.brk.len}{only combine segments where the distance between the two is less than max.brk.len (in mbp)}

\item{ap.calc}{set to true if calculating aneuploidy, which uses different criteria than HRD}
}
\value{
a truncated data frame with the combined segments
}
\description{
Combine two chromosomal segments into one
}
\details{
For HRD: if CNt, A, and B are equal in two adjacent segments, and break length < max.brk.len, combine
the two segments into one. For aneuploidy; if aneuploidy status is equal in two adjacent segments, combine into one.
}
\examples{
## There are 4 segments on chromosome 14
seq.dat <- sub01.segments[ sub01.segments$chromosome == "chr14",]
seq.dat

## The gap between segments 3 and 4 is very small; copynumber, A, and B are identical in both segments
## this is a false split. It will be combined into one segment. 
seq.dat <- combineSeg(seq.dat)
seq.dat

## the data now has 3 segments.
}
