% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAneuploidy.R
\name{getAneuploidy}
\alias{getAneuploidy}
\title{Compute the aneuploidy score for a given chromosome}
\usage{
getAneuploidy(seq.dat, ploidy.dat, chr, ref = "grch37")
}
\arguments{
\item{seq.dat}{the data.frame of sequencing data}

\item{ploidy.dat}{the data.frame of ploidy data}

\item{chr}{the chromosome of interest}
}
\description{
Compute the aneuploidy score for a given chromosome
}
\details{
Aneuploidy status is defined as amplified if main copynumber > ploidy; neutral if main copy number and ploidy
are equal; and deleted if main copy number < ploidy. Segments are combined if the aneuploidy status of two adjacent
segments is equal, and the segments do not cross the centromere.
}
\seealso{
\code{\link{combineSeg}}
}
