% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAneuploidy.R
\name{getAneuploidyGenome}
\alias{getAneuploidyGenome}
\title{get aneuploidy genome}
\usage{
getAneuploidyGenome(seq.dat, ploidy.dat, include.X = FALSE)
}
\arguments{
\item{seq.dat}{the data.frame of sequencing data}

\item{ploidy.dat}{the data.frame of ploidy data}

\item{include.X, }{include chromosome (logical}
}
\description{
get aneuploidy genome
}
\details{
get aneuploidy status across the whole genome
}
\seealso{
\code{\link{getAneuploidy}}
}
