% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAneuploidy.R
\name{getAneuploidyScore}
\alias{getAneuploidyScore}
\title{get aneuploidy score}
\usage{
getAneuploidyScore(dat)
}
\arguments{
\item{dat, }{data.frame containing aneuploidy scores}
}
\description{
get aneuploidy score
}
\details{
description goes here
}
\seealso{
\code{\link{getAneuploidy}}
}
