% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNT.R
\name{getCNt}
\alias{getCNt}
\title{get the main copynumber state for each chromosome}
\usage{
getCNt(seq.dat)
}
\arguments{
\item{seq.dat}{the data.frame of sequencing data}
}
\value{
a data.frame containing the main copynumber state for each input chromosome
}
\description{
get the main copynumber state for each chromosome
}
\details{
Get the main copynumber state for each chromosome. Calculate the total proportion of
the chromosome associated with each unique copynumber state; return the copynumber state that is most
represented.
}
\examples{
seq.dat <- preprocessHRD( seq.dat )
CN.dat <- getCNt( seq.dat )

}
