% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLOH.R
\name{getLOH.raw}
\alias{getLOH.raw}
\title{Compute the number of loss of heterozygosity (LOH) events}
\usage{
getLOH.raw(seq.dat, min.seg.size = 1.5e+07)
}
\arguments{
\item{seq.dat}{a data.frame of the sequencing data}

\item{min.seg.size}{minimum size of the seqgment}

\item{ploidy}{the estimated ploidy}
}
\description{
Compute the number of loss of heterozygosity (LOH) events
}
\details{
raw LOH is calculated as the number of events where: the segment length is greater than the minimum segment size 
(default of 15e06), and the segment size is less than 90% of the length of the entire chromosome.
Normalized LOH is calculated as raw LOH * k, the ploidy correction factor
}
\seealso{
\code{\link{ploidyCorrectionFactor}}
}
