% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNTAI.R
\name{getNTAI.raw}
\alias{getNTAI.raw}
\title{Compute the number of non-telomeric allelic imbalance (NTAI) events}
\usage{
getNTAI.raw(seq.dat, min.seg.size = 1.1e+07)
}
\arguments{
\item{seq.dat}{the data.frame of sequencing data}

\item{min.seg.size}{the minimum segment size}

\item{CN.dat}{the data.frame of copy number information}
}
\value{
the number of NTAI events
}
\description{
Compute the number of non-telomeric allelic imbalance (NTAI) events
}
\details{
raw NTAI is calculated as the number of segments where the segment size is greater than the 
minimum segment size; allelic imbalance is present; the segment does not cross the centromere; and the
segment is not in either of the telomeres. NTAI is normalized by removing all main copy number segments.
}
\examples{
seq.dat <- preprocessSeq( seq.dat )
CN.dat <- getCNt( seq.dat )

## the number of NTAI events in seq.dat
ntai <- getNTAI.norm( seq.dat, CN.dat)

}
