% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrd.stats.R
\name{hrd.stats}
\alias{hrd.stats}
\title{get hrd metrics}
\usage{
hrd.stats(seq.dat, CN.dat, ref = "grch37")
}
\arguments{
\item{seq.dat}{the preprocessed sequencing data}

\item{CN.dat}{the copy number data}

\item{scaleTotal}{optional: rescale the total to range of 0-100}

\item{type}{optional: type of score to return, one of "sum" or "average"}
}
\value{
LOH, NTAIraw, NTAInorm, LST, total
}
\description{
get hrd metrics
}
\examples{
seq.dat <- hrd.stats(  )

}
