% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inCentromere.R
\name{in.centromere}
\alias{in.centromere}
\title{Return an index of segments that are in the centromere, for a given chromosome}
\usage{
in.centromere(start.pos, end.pos, ref.cent.start, ref.cent.end)
}
\arguments{
\item{start.pos}{start position of the segment}

\item{end.pos}{end position of the segment}

\item{ref.cent.start}{start position of the centromere}

\item{ref.cent.end}{end position of the centromere}
}
\value{
logical index of segments crossing the centromere
}
\description{
Return an index of segments that are in the centromere, for a given chromosome
}
\details{
This function returns a logical vector with value of TRUE if the corresponding segment crosses
the centromere.
}
