% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyNorm.log.R
\name{ploidyNorm.log}
\alias{ploidyNorm.log}
\title{Correct HRD metric for ploidy assuming a log-linear relationship}
\usage{
ploidyNorm.log(HRD, ploidy)
}
\arguments{
\item{HRD}{vector of numeric values, the HRD metric to normalize. Can be HRD.Score or any individual measure.}

\item{ploidy}{vector of ploidy values.}
}
\value{
logHRD corrected for ploidy
}
\description{
Correct HRD metric for ploidy assuming a log-linear relationship
}
\details{
Regress out the effect of ploidy from HRD
}
