% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessHRD.R
\name{preprocessHRD}
\alias{preprocessHRD}
\title{Define additional properties of the sequencing data necessary to compute HRD}
\usage{
preprocessHRD(seq.dat, ref)
}
\arguments{
\item{seq.dat}{the raw sequencing data}

\item{ref}{the reference genome to use (grch37, grch38)}
}
\value{
seq.dat
}
\description{
Define additional properties of the sequencing data necessary to compute HRD
}
\details{
define chromosome size, arms, telomere position, and centromere position from reference data; 
compute allelic imbalance and segment size
}
\examples{
seq.dat <- sub01.segments[ sub01.segments$chromosome == "chr1",]

seq.dat <- preprocessHRD( seq.dat )
seq.dat
}
