% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSeq.R
\name{preprocessSeq}
\alias{preprocessSeq}
\title{Define additional properties of the sequencing data necessary to compute HRD}
\usage{
preprocessSeq(seq.dat, ref = "grch37")
}
\arguments{
\item{seq.dat}{the raw sequencing data}
}
\value{
seq.dat
}
\description{
Define additional properties of the sequencing data necessary to compute HRD
}
\details{
define chromosome size, arms, telomere position, and centromere position from reference data; 
compute allelic imbalance and segment size
}
\examples{
seq.dat <- sub01.segments[ sub01.segments$chromosome == "chr1",]

seq.dat <- preprocessSeq( seq.dat )
seq.dat
}
